test for en and fr in reportersList, evelopeDetail, obligationDetail
