+++ frontend/assets/js/main.js
@@ -6,14 +6,15 @@ import App from './App';
 import router from './router';
 import '../css/main.scss';
 import VueCookies from 'vue-cookies'
-import ChannelsPlugin from './mixins/ChannelsPlugin';
+import ChannelsPlugin from './setup/ChannelsPlugin';
+import { i18n } from './setup/i18n-setup';
 
-const channelOptions = { debug: true, reconnectInterval: 3000 }
-// Globally register bootstrap-vue components
+const channelOptions = { debug: true, reconnectInterval: 3000 };
+
 Vue.use(BootstrapVue);
 Vue.use(ChannelsPlugin, channelOptions);
-// Globally register vue cookies
-Vue.use(VueCookies)
+Vue.use(VueCookies);
+
 Vue.config.productionTip = false;
 
 /* eslint-disable no-new */
@@ -22,4 +23,5 @@ new Vue({
   router,
   template: '<App/>',
   components: { App },
+  i18n
 });
