+++ frontend/assets/js/setup/ChannelsPlugin.js
