+++ package.json
@@ -27,6 +27,7 @@
     "vue": "^2.5.2",
     "vue-backtotop": "^1.3.9",
     "vue-cookies": "^1.5.6",
+    "vue-i18n": "^7.6.0",
     "vue-router": "^3.0.1",
     "webpack-bundle-tracker": "^0.2.1"
   },
